Deprecate PreloadHeroImage for OptimizeHeroImages
