+++ src/Optimizer/DefaultConfiguration.php
@@ -4,6 +4,7 @@ namespace AmpProject\Optimizer;
 
 use AmpProject\Optimizer\Configuration\AmpRuntimeCssConfiguration;
 use AmpProject\Optimizer\Configuration\OptimizeAmpBindConfiguration;
+use AmpProject\Optimizer\Configuration\OptimizeHeroImagesConfiguration;
 use AmpProject\Optimizer\Configuration\PreloadHeroImageConfiguration;
 use AmpProject\Optimizer\Configuration\RewriteAmpUrlsConfiguration;
 use AmpProject\Optimizer\Configuration\TransformedIdentifierConfiguration;
@@ -12,6 +13,7 @@ use AmpProject\Optimizer\Exception\UnknownConfigurationClass;
 use AmpProject\Optimizer\Exception\UnknownConfigurationKey;
 use AmpProject\Optimizer\Transformer\AmpRuntimeCss;
 use AmpProject\Optimizer\Transformer\OptimizeAmpBind;
+use AmpProject\Optimizer\Transformer\OptimizeHeroImages;
 use AmpProject\Optimizer\Transformer\PreloadHeroImage;
 use AmpProject\Optimizer\Transformer\RewriteAmpUrls;
 use AmpProject\Optimizer\Transformer\TransformedIdentifier;
@@ -43,6 +45,7 @@ class DefaultConfiguration implements Configuration
     protected $transformerConfigurationClasses = [
         AmpRuntimeCss::class         => AmpRuntimeCssConfiguration::class,
         OptimizeAmpBind::class       => OptimizeAmpBindConfiguration::class,
+        OptimizeHeroImages::class    => OptimizeHeroImagesConfiguration::class,
         PreloadHeroImage::class      => PreloadHeroImageConfiguration::class,
         RewriteAmpUrls::class        => RewriteAmpUrlsConfiguration::class,
         TransformedIdentifier::class => TransformedIdentifierConfiguration::class,
